make favoriteExercises show up on calendar
