+++ components/FormStrengthAndRunning/index.js
@@ -15,6 +15,7 @@ export default function FormStrengthAndRunning({
   handleChange,
   inputText,
   day,
+  favoriteExercises,
 }) {
   return (
     <>
@@ -72,6 +73,7 @@ export default function FormStrengthAndRunning({
 
         {sportSelected === "strength" ? (
           <StrengthForm
+            favoriteExercises={favoriteExercises}
             handleSubmit={handleSubmit}
             handleCancelClick={handleCancelClick}
             handleChange={handleChange}
@@ -80,6 +82,7 @@ export default function FormStrengthAndRunning({
           />
         ) : (
           <RunningForm
+            favoriteExercises={favoriteExercises}
             handleSubmit={handleSubmit}
             handleCancelClick={handleCancelClick}
             handleChange={handleChange}
